+++ packages/monaco-graphql-editor/src/index.tsx
@@ -1,0 +1,3 @@
+export * from './editor/SchemaEditor';
+export * from './editor/utils';
+export * from './editor/EnrichedLanguageService';