+++ packages/the-guild-components/CHANGELOG.md
@@ -1,6 +1,55 @@
 # the-guild-components
 
+## 1.3.1
+
+### Patch Changes
+
+- Updated dependencies [1a290f0]
+  - @theguild/components@1.3.1
+
+## 1.3.0
+
+### Minor Changes
+
+- 13b1df6: extended footer component + storybook deprecation fixes
+
+### Patch Changes
+
+- Updated dependencies [2aa293d]
+- Updated dependencies [13b1df6]
+  - @theguild/components@1.3.0
+
+## 1.2.4
+
+### Patch Changes
+
+- Updated dependencies [c256e07]
+  - @theguild/components@1.2.4
+
+## 1.2.3
+
+### Patch Changes
+
+- Updated dependencies [b68016b]
+- Updated dependencies [9c5cd85]
+  - @theguild/components@1.2.3
+
+## 1.2.2
+
+### Patch Changes
+
+- Updated dependencies [d2257c4]
+  - @theguild/components@1.2.2
+
+## 1.2.1
+
+### Patch Changes
+
+- Updated dependencies [1a94a7a]
+  - @theguild/components@1.2.1
+
 ## 1.2.0
+
 ### Minor Changes
 
 - aecc55f: included envelop components & fixes for ssr